* making our DynamicData package more typesafe. the size of the diff might be scaring, but most of it is removal of type-casts or changing how Key is constructed.   * modified the Key class to use a generic type <T>   * modified the KeyedMap to make use o this generic type and add type-safety to it's operations, letting the compiler tell the errors (avoiding runtime errors).   * look at the diff carefully, check out if I missed something.  * WARNING! user files, vfs metadata (xmls) or any other JavaBean that serializes through XMLDecoder/XMLEncoder must be recreated since the Key constructor has changed (I had to change the DefaultPersistenceDelegate)
 * refactored some jprintf() methods from Session and BaseFtpConn and removed some the unnused ones

+++ src/plugins/org.drftpd.usermanager.javabeans/src/org/drftpd/usermanager/javabeans/BeanUser.java
@@ -94,7 +94,7 @@ public class BeanUser extends AbstractUser {
 	 * Setter for userfile backwards comptibility. Should work but i had nothing
 	 * to test with.
 	 */
-	public void setGroupLeechSlots(short s) {
+	public void setGroupLeechSlots(int s) {
 		getKeyedMap().setObject(UserManagement.LEECHSLOTS, s);
 	}
